+++ AHM2_ch03/AHM2_03.4.1-3.R
@@ -11,7 +11,6 @@
 
 library(AHMbook)
 library(jagsUI)
-bugs.dir <- "C:/WinBUGS14" # the location of the WinBUGS14.exe file on your machine
 
 # 3.4 Spatial hierarchical CJS models
 # ===================================